#128: THE POKES ASSUMED AN ENGINE THAT DECIDES — 21 lanes re-grounded, the drivers fixed, and one TS refusal that diverged the engines
The seed lock now checks BEFORE seed rewrites it (a check after diffs a
generation against itself and can never fail).

Every red GPU poke lane is green. None was a live-path mechanic bug; the
shapes, recorded in docs/AUDIT.md: the auto-decision premise (buys, strikes
and spreads are wire ORDERS now — order helpers added), the registry
confound (city_dist_tile, twice more), the dead civ-index/pool-slot
conventions, the wrong resolver, a poked plane under an unbumped
_eff_version cache, and purged alias names still asserted by name.

The REAL finds were all on the driven path the warm-up opened:
- replay_seat shipped 0-dim tech/civic tensors (the record apply gathers on
  dim 1);
- _decide_turn dropped production_tile from its own apply, so the drive and
  its own record built different worlds;
- the virtual planner unpacked a _war_march_target signature that no longer
  exists;
- the driver ordered FOUND unconditionally (the mask column is type-only),
  pinning any settler off legal ground to a refused verb forever — there
  were no settle targets at all. _settle_targets + a rank-0 civilian walk
  toward job/spread/settle destinations fixed both; barbs still eat lone
  settlers, which is faithful, so lanes needing a second city use the new
  plant_city (the engine's own spawn + FOUND over the order path);
- an f32 sim crashed on three f64 tensor-index writes in the loyalty,
  growth and produce bodies (f64 intermediates now stored at plane dtype);
- TS meleeAttackInner refused a barbarian at a civ city for peace its own
  targeting predicate had waived — the GPU sacks, so the engines answered
  differently. The refusal now mirrors the targeting predicate.

drive_test asserts absolute floors from the settler start (the scripted
reference it compared against was a seat that does nothing). warmup.py
gains developed() (ladder-driven worlds) and plant_city().

+++ gpu/core/sim_phase.py
@@ -238,7 +238,10 @@ class SimPhase:
         upd = act & others
         nxt = torch.where(upd, (loy + delta).clamp(min=0, max=lmax), loy)
         cap = self.city_is_cap[bidx, row, col]
-        self.city_loyalty[bidx, row, col] = torch.where(upd & cap, torch.full_like(nxt, lmax), nxt)
+        # f64 intermediates, stored at the PLANE's dtype (an f32 sim keeps an
+        # f32 loyalty plane).
+        self.city_loyalty[bidx, row, col] = torch.where(
+            upd & cap, torch.full_like(nxt, lmax), nxt).to(self.city_loyalty.dtype)
         return upd & ~cap & (self.city_loyalty[bidx, row, col] <= 0)
 
     def _seat_loyalty_flips(self, row: int, flip: torch.Tensor) -> None:
@@ -282,7 +285,8 @@ class SimPhase:
         grow = act & (box >= need)
         starve = act & ~grow & (box < 0)
         nxt = torch.where(grow, box - need, torch.where(starve, torch.zeros_like(box), box))
-        self.city_growth[bidx, row, col] = torch.where(act, nxt, old)
+        # f64 intermediates, stored at the PLANE's dtype (see _seat_city_loyalty)
+        self.city_growth[bidx, row, col] = torch.where(act, nxt, old).to(old.dtype)
         pop = self.city_pop[bidx, row, col] + grow.long()
         self.city_pop[bidx, row, col] = torch.where(starve, (pop - 1).clamp(min=1), pop)
 
@@ -318,7 +322,8 @@ class SimPhase:
         # phase.ts spends the bank right after the production add.
         prog = self.city_progress[bidx, row, col]
         bank = self.city_prod_bank[bidx, row, col]
-        self.city_progress[bidx, row, col] = torch.where(has_q, prog + prod + bank, prog)
+        # f64 intermediates, stored at the PLANE's dtype (see _seat_city_loyalty)
+        self.city_progress[bidx, row, col] = torch.where(has_q, prog + prod + bank, prog).to(prog.dtype)
         self.city_prod_bank[bidx, row, col] = torch.where(has_q, torch.zeros_like(bank), bank)
         cost = self.city_cost[bidx, row, col].clone()
         done = has_q & (self.city_progress[bidx, row, col] >= cost)
